Created the game_files.py
game_files.py will create the csv files for the round information that the user can input and the processed round information where the net_profit will be after all the calculations are applied. Also added in a call in the index method to generate these files when the index page is loaded. This will be removed in future but for now its for testing purposes