Rounds can now be entered and processed
The user can now enter round information and the round will be processed. This processing will, check to see if the banker has been played, steal money if the banker has been played, calc the net profit, calc the penalties, cald the best peddle in the players hand to remove from the net profit, calc the bonus for the player with the highest net profit, wrie the information to the processed round file, update the round round to keep track of what round was played.